improve status check
